trim pinned routing fallout
