+++ lib/bindings/python/rust/llm/kv.rs
@@ -1077,6 +1077,7 @@ impl KvRouter {
                     lora_name.clone(),
                     0.0,
                     None,
+                    None,
                     None, // allowed_worker_ids: pass via RoutingHints in PreprocessedRequest path
                 )
                 .await
